Moved reperesentations and request parameters from Operation module
