+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/representation/SettingRepresentation.java
@@ -7,12 +7,10 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.rest.operation.data;
+package org.oscm.rest.common.representation;
 
 import org.oscm.internal.types.enumtypes.ConfigurationKey;
 import org.oscm.internal.vo.VOConfigurationSetting;
-import org.oscm.rest.common.representation.Representation;
-import org.oscm.rest.common.representation.RepresentationCollection;
 
 import javax.ws.rs.WebApplicationException;
 import java.util.ArrayList;
