+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/requestparameters/OperationParameters.java
@@ -7,7 +7,7 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.rest.operation;
+package org.oscm.rest.common.requestparameters;
 
 import org.oscm.rest.common.requestparameters.RequestParameters;
